Issue #40 Update menu config menu_item_tiles to unsigned short.
There was an apparently memory limitation stopping the original array
size.
An unsigned short will be plenty since we only have 128 tiles and
7 palettes.
This now fixes previous issues with displaying menu tiles.
Issue #34
-e 
Original/build commit ID: b5c116f1ffb1b6aabc55ef11f692ca6223cc50fa

+++ menu_config.h
@@ -46,7 +46,7 @@
 
 typedef struct {
     // Menu can contain 8 items, split into tile data and palette data, each with 12 tiles
-    const unsigned int menu_item_tiles[8][2][12];
+    const unsigned short menu_item_tiles[8][2][12];
 
     // Bit mask of menu items. Lower nibble is left column (x = 0), upper nibble is right (x = 1)
     const UINT8 menu_items;
